also run unit tests as root
