+++ sys_conn_test.go
@@ -2,8 +2,6 @@ package quic
 
 import (
 	"net"
-	"os"
-	"runtime"
 	"time"
 
 	"github.com/quic-go/quic-go/internal/protocol"
@@ -31,32 +29,5 @@ var _ = Describe("Basic Conn Test", func() {
 		Expect(p.data).To(Equal([]byte("foobar")))
 		Expect(p.rcvTime).To(BeTemporally("~", time.Now(), scaleDuration(100*time.Millisecond)))
 		Expect(p.remoteAddr).To(Equal(addr))
-	})
-})
-
-var _ = Describe("Can change the receive buffer size", func() {
-	It("Force a change (if we have CAP_NET_ADMIN)", func() {
-		if runtime.GOOS != "linux" {
-			Skip("Only an option on linux")
-		}
-
-		if os.Getuid() != 0 {
-			Skip("Must be root to force change the receive buffer size")
-		}
-
-		c, err := net.ListenPacket("udp", "127.0.0.1:0")
-		Expect(err).ToNot(HaveOccurred())
-		forceSetReceiveBuffer(c, 256<<10)
-
-		size, err := inspectReadBuffer(c)
-		Expect(err).ToNot(HaveOccurred())
-		//  The kernel doubles this value (to allow space for bookkeeping overhead)
-		Expect(size).To(Equal(512 << 10))
-
-		forceSetReceiveBuffer(c, 512<<10)
-		size, err = inspectReadBuffer(c)
-		Expect(err).ToNot(HaveOccurred())
-		//  The kernel doubles this value (to allow space for bookkeeping overhead)
-		Expect(size).To(Equal(1024 << 10))
 	})
 })
